+++ src/components/IssueSearchAdvanced.vue
@@ -28,10 +28,11 @@ const nameForOption = (fieldName: string) => {
     acknowledgments__name: 'Acknowledgment Author',
     affects__trackers__errata__advisory_name: 'Errata Advisory Name',
     affects__trackers__external_system_id: 'Tracker External System ID',
-    workflow_state: 'Flaw Status',
+    workflow_state: 'Flaw State',
     cwe_id: 'CWE ID',
     cve_id: 'CVE ID',
     source: 'CVE Source',
+    cve_description: 'CVE Description',
     requires_cve_description: 'CVE Description Review',
     major_incident_state: 'Incident State',
   };