+++ src/components/widgets/EditableDate.vue
@@ -6,7 +6,7 @@
 // Pressing escape or clicking the abort button aborts the change
 
 import { DateTime } from 'luxon';
-import { ref, nextTick, reactive, toValue } from 'vue';
+import { ref, nextTick, reactive, watch } from 'vue';
 import { IMask } from 'vue-imask';
 
 const props = defineProps<{
@@ -16,18 +16,17 @@ const props = defineProps<{
   editing?: boolean,
   error?: string | null,
 }>();
-const initialValue = toValue(props.modelValue);
+
 const emit = defineEmits<{
   'update:modelValue': [value: string | undefined],
   'update:editing': [value: boolean],
 }>();
 
-
 const elInput = ref<HTMLInputElement>();
 const elDiv = ref<HTMLDivElement>();
 const editing = ref<boolean>(props.editing ?? false);
 
-const pattern = props.includesTime ? 'yyyy-MM-dd hh:mm' : 'yyyy-MM-dd';
+const pattern = props.includesTime ? 'yyyy-MM-dd hh:mm Z' : 'yyyy-MM-dd';
 const maskLayer = {
   mask: Date,
   pattern,
@@ -65,7 +64,10 @@ const maskLayer = {
       from: 0,
       to: 59,
       maxLength: 2,
-    }
+    },
+    Z: {
+      mask: 'UTC',
+    },
   },
   // IMask's complete event depends on format and parse returning valid dates
   format: formatDate,
@@ -73,7 +75,7 @@ const maskLayer = {
   autofix: true,
 };
 
-const formatString = props.includesTime ? 'yyyy-MM-dd T' : 'yyyy-MM-dd';
+const formatString = props.includesTime ? 'yyyy-MM-dd T\' UTC\'' : 'yyyy-MM-dd';
 // The assumption is that the Date or String is in UTC
 function formatDate(date: Date | string): string {
   const jsDate = new Date(date); // Handles strings in ISO/component format, and Date object
@@ -81,12 +83,14 @@ function formatDate(date: Date | string): string {
 }
 // The assumption is that the Date or String is in UTC
 function parseDate(dateString: string): Date {
-  return DateTime.fromFormat(dateString, formatString, { zone: 'utc' }).toJSDate();
+  // Remove the static part ' UTC' before parsing
+  const cleanedDateString = dateString.replace(' UTC', '').trim();
+  return DateTime.fromFormat(cleanedDateString, formatString.replace('\' UTC\'', ''), { zone: 'utc' }).toJSDate();
 }
-
 const maskState = reactive({
   completed: false,
-  masked: '',
+  // Force caret to start of input
+  masked: ' ',
   unmasked: '',
 });
 
@@ -143,8 +147,8 @@ function commit() {
 
 function abort() {
   editing.value = false;
-  maskState.masked = initialValue || '';
-  maskState.unmasked = initialValue || '';
+  maskState.masked = props.modelValue || '';
+  maskState.unmasked = props.modelValue || '';
 }
 
 function onBlur(e: FocusEvent | null) {
@@ -158,7 +162,6 @@ function onBlur(e: FocusEvent | null) {
     return;
   }
 
-
   if (e.relatedTarget instanceof Node) {
     if (elDiv.value?.contains(e.relatedTarget)) {
       return; // Don't abort or commit if focus is still within the input editing parts of the component
@@ -168,7 +171,6 @@ function onBlur(e: FocusEvent | null) {
     }
   }
 }
-
 
 function isValidDate(d: Date | string): boolean {
   if (typeof d === 'string') {
@@ -189,7 +191,9 @@ function osimFormatDate(date?: string | null): string {
   return formattedDate;
 }
 
-
+watch(() => props.modelValue, () => {
+  maskState.masked = props.modelValue || '';
+});
 
 </script>
 
@@ -217,7 +221,7 @@ function osimFormatDate(date?: string | null): string {
     </div>
 
     <div
-      v-if="editing"
+      v-show="editing"
       ref="elDiv"
       class="input-group osim-date-edit-field"
       @blur="onBlur($event)"